pyproject.toml
[project]
name = "mercadoPagoDjango"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "jonatasperaza", email = "jonatassilvaperaza@gmail.com"},
]
dependencies = [
    "django>=5.0.7",
    "mercadopago>=2.2.3",
    "djangorestframework>=3.15.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true

[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
